Tried to insert C and failed; know why.
When we link in the C code, it goes below the magic number,
which means that when we try and jump we jump into memory
that hasn't been loaded in yet. We can mitigate this by
changing up the linker script a bit. This should also have
the side effect of not including those weird buffer sections,
cause we can just include that in the linker script.